+++ feature/widget/src/main/res/drawable/ic_widget_refresh.xml
@@ -1,0 +1,20 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M16.667,3.333v5h-5"
+      android:strokeWidth="1.667"
+      android:strokeColor="#5B5F72"
+      android:strokeLineCap="round"
+      android:strokeLineJoin="round"
+      android:fillColor="#00000000"/>
+  <path
+      android:pathData="M14.767,12.5a5.833,5.833 0,1 1,-1.35,-6.067L16.667,8.333"
+      android:strokeWidth="1.667"
+      android:strokeColor="#5B5F72"
+      android:strokeLineCap="round"
+      android:strokeLineJoin="round"
+      android:fillColor="#00000000"/>
+</vector>
